refactor: `Header`의 필터 선택 관련 기능을 `Main` 스크린으로 분리 - 필터 선택 버튼에 드롭다운 화살표 아이콘 추가
