+++ src/components/Header/Header.styles.js
@@ -16,6 +16,8 @@ Container.Top = styled.View`
   justify-content: space-between;
 `;
 
+Container.Bottom = styled.View``;
+
 export const IconButton = styled.TouchableOpacity`
   padding: 16px;
 `;
@@ -23,23 +25,4 @@ export const IconButton = styled.TouchableOpacity`
 export const RightSide = styled.View`
   flex-direction: row;
   align-items: center;
-`;
-
-export const FilterSelect = styled.View`
-  width: 100%;
-  flex-direction: row;
-  justify-content: space-between;
-  padding: 8px 34px 34px 34px;
-`;
-
-FilterSelect.Item = styled.TouchableOpacity`
-  margin-right: 8px;
-  padding: 12px 16px;
-  border-radius: 20px;
-  border: solid 1px ${({ active }) => (active ? '#ff9d33' : '#f0f0f0')};
-`;
-
-FilterSelect.Text = styled.Text`
-  font-size: 12px;
-  font-weight: ${({ active }) => (active ? 'bold' : 'normal')};
 `;
